Finished a working search system.
Modified the header to show search results
Added board_id to the list of things sent back by search.php
Added search events and displaying of results (with ajax request)
Added options node_id and sub_id in the query (after ...php? in the url) and the board now moves to center the searched node or subtitle and shows its content.
Added a bunch of css code for search results
Board.js was modified in order to select the searched element once the board is fully loaded.